Software Concurrent Scavenge Support on Power
Sets OMR_GC_CONCURRENT_SCAVENGER in the configure files to enable
concurrent scavenge.

Modifies mminit.cpp to be able to set softwareRangeCheckReadBarrier to
true to enable software based concurrent scavenge.

Implemented irdbarEvaluator and ardbarEvaluator which are used to handle
loads that might need to call out to the readbarrier under Concurrent
Scavenge GC policy.

Updated VMinlineCompareAndSwapObject to be aware of the possiblity of
calling out to the read barrier under Concurrent Scavenge GC policy.

Implemented a new version of arraycopyEvaluator for reference array copies
that calls out to the referenceArrayCopy if current scavenge is currently
happening.

Added jitReadBarrier and jitReferenceArrayCopy as helpers to call out to
J9ReadBarrier and referenceArrayCopy.

Updated thisThreadGetConcurrentScavengeActiveByteAddressOffset and
thisThreadGetEvacuateBaseAddressOffset to recognize Power as well as X86.

Updated insertPrefetchIfNecessary to recognize the irdbari opcode.

Updated copyright dates on modified files.

TM is current not compatible with read barriers. This change will disable
TM if read barriers are needed.

Issue: #3421
Signed-off-by: jimmyk <jimmyk@ca.ibm.com>