Avoid unnecessary SVM records for field references
Certain records generated by the symbol validation manager (SVM)
identify the class in which an (instance or static) field was found,
based on a constant pool field reference. If the beholder refers to one
of its own fields, or to a field of a well-known class, and if it does
so by naming the class that declares the field (not a subclass), then no
record is necessary. The class reference is unambiguous and it must
succeed, and because the class chain of the named class will have been
validated, the field must be found again directly in that class.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>