Move scavenger functionality from CLI to new delegate
The scavenger callbacks in the CLI are being replaced with a new
ScavengerDelegate glue class. The code is already in place in OMR, under
the ifdef'd feature OMR_GC_SCAVENGER_DELEGATE.

This commit moves the CLI code in java to the new scavenger delegate.
This is a mechanical move/rename PR.

Summary of changes:

- Enable the `OMR_GC_SCAVENGER_DELEGATE` flag in J9
- Move implementation of scavenger-related CLIJ callbacks to ScavengerDelegate
  - callbacks no longer have `scavenger_` prefix
  - callbacks aren't virtual
- Move root scavenging flags from CLI to scavenger
  - EG `_shouldScavengeFinalizableObjects`
  - flags no longer have `scavenger_` prefix
  - move flag accessors from CLIJ to delegate
  - In ScavengerRoot{Scanner, Clearer} access flags through scavenger()->getDelegate().
- Remove J9 VM and OMR VM fields from CLIJ, they're no longer in use.
- Make delegate a friend of ObjectAccessBarrier, not CLIJ.

Signed-off-by: Robert Young <rwy0717@gmail.com>