Fix Java12 test compilation
In Java 12 internal libraries SharedSecrets and JavaLangAccess were moved from jdk.internal.misc to jdk.internal.access causing some tests in Java8AndUp to fail.

To reuse as little code as possible I've moved all the tests that were in src_110 to a folder called src_110_up which will be used for versions 11+. I've wrapped the uses of SharedSecrets and JavaLangAccess into a class called InternalAccessor. There is one InternalAccessor for Java 11 in src_110 and one for Java 12 in src_120.

Signed-off-by: Theresa Mammarella <Theresa.T.Mammarella@ibm.com>