Fix x86 PicBuilder VPIC data layout assumptions
PicBuilder on x86 was modified in bada7c8 to locate fields of the VPIC
data using named constants rather than hard-coded offsets, but there
were a few remaining places where the code continued to rely on implicit
assumptions about the layout. In particular,

- cpAddr was assumed to have offset 0, with pointers to the beginning of
  the VPIC data being treated as pointers to cpAddr; and

- on 32-bit, cmpRegImm4ModRM was assumed to be the last byte.

PicBuilder is updated to avoid relying on these implicit assumptions, in
order to facilitate the rearrangement of the fields of the VPIC data.

Because of jitResolveVirtualMethod(), cpIndex must still always follow
immediately after cpAddr.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>