Align mutable fields in x86 IPIC data
Because the interface class and itable offset can be written at runtime,
and might be read concurrently by another thread, they need to be
aligned so that the accesses will be atomic. Currently it's possible to
observe tearing, which can lead PicBuilder to attempt direct dispatch
with a bogus J9Method pointer. Such tearing might also interfere with
standard interface dispatch.

The start of the snippet is now adjusted to effect the alignment.

Fixes #4446
Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>